requirements.txt
azure-functions
mysql-connector-python
python-dotenv
auth0-python
python-jose[cryptography]